ppa: Hopefully fix libva/libbluray
This is a seriously nasty fix, so please let me know if you've got a nicer way.
Untested but I believe it should work.

The substitution for var_info is run on the user's (buildbot in this case),
machine, not the ppa-builder. So lsb-release reports the version installed,
not the one that's being built.

This could probably be grep'd from a changelog or similar, but I could not find
a good tool for the job.

I repeat, hack hack hack!